odyssey-mcp-server/requirements.txt
# MCP Server Dependencies
fastmcp>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0.0
pathlib>=1.0.1

# Code Analysis
tree-sitter>=0.21.0
regex>=2023.0.0

# Utilities
python-dotenv>=1.0.0
